refactor: schema generator
