+++ packages/amplify-graphql-schema-generator/src/index.ts
@@ -2,3 +2,5 @@ export * from './datasource-adapter';
 export * from './schema-generator';
 export * from './utils';
 export * from './schema-representation';
+export * from './generate';
+export * from './input';
